Added Alloy skeleton I tried setting up a directory structure that made sense (facts, preds etc in seperate directories) but I couldn't get Alloy to open files that were in folders up (.. creates a syntax error). Furthermore, I couldn't find how to use other modules (I hope I will find out this weekend) like this:
file1:
	open def1
	open use1

def1:
	sig definition {}

use1:
	sig usage{
		def : definition
	}

This causes a syntax error in use1.als because definition isn't defined.
I had hoped to be able to use this to open dummy files where we don't
need the full functionality and the normal files where we do.
This is not working so far so you have to change the open statements
in individual files if you want to change something.

The actual checking and visualization are in the pred*.als files.
I have one non-trivial predicate working, namely userAgreed. When
run, this actually generates two instances of the Vote class for
which an instance of the User class doesn't agree at first, but
does later.
The other pred file, predShowCore, shows an instance of the Core,
User, Event and Reponse. If you project over User or Event, you can see
that this is working okay for now.